refactor: change modules name
fix: update neovim removing deprecated modules
