+++ package.json
@@ -4,11 +4,11 @@
   "main": "index.js",
   "license": "MIT",
   "scripts": {
-    "dev": "nodemon ts-node src/signup3.ts",
+    "dev": "nodemon ts-node src/main.ts",
     "docker:start": "docker compose up -d",
     "docker:stop": "docker compose stop",
     "docker:clean": "docker compose rm -fsv",
-    "coverage": "npx nyc --reporter=lcov --reporter=text ts-node src/signup.ts"
+    "coverage": "npx nyc --reporter=lcov --reporter=text ts-node src/main.ts"
   },
   "dependencies": {
     "@types/express": "^4.17.21",
