+++ test/unit/Circle.test.ts
@@ -1,0 +1,6 @@
+import Circle from "../../src/Circle";
+
+test("Should calculate the area of circle", function () {
+    const circle = new Circle(2);
+    expect(circle.getArea()).toBe(12.566370614359172);
+});
