+++ test/unit/validateCpf.test.ts
@@ -1,4 +1,4 @@
-import { validateCpf } from "../src/validateCpf";
+import { validateCpf } from "../../src/validateCpf";
 
 test.each([
     "97456321558",
